requirements.txt
streamlit==1.27.0
tensorflow==2.20.0
pillow==10.0.1
numpy==1.25.0
pandas==2.1.0
